+++ tramontane/memory/longterm.py
@@ -58,7 +58,7 @@ class LongTermMemory:
         if self._db is not None:
             return self._db
 
-        self._db = sqlite3.connect(self._db_path)
+        self._db = sqlite3.connect(self._db_path, check_same_thread=False)
         self._db.row_factory = sqlite3.Row
         self._db.execute("PRAGMA journal_mode=WAL")
         self._db.execute("PRAGMA foreign_keys=ON")